requirements.txt
django-cors-headers==3.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
